+++ tests/integration/high_availability/test_upgrade_from_stable.py
@@ -1,0 +1,75 @@
+# Copyright 2023 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+import asyncio
+import logging
+import pytest
+
+from integration.high_availability.high_availability_helpers import (
+    ensure_all_units_continuous_writes_incrementing,
+    relate_mysql_and_application,
+)
+from pytest_operator.plugin import OpsTest
+
+logger = logging.getLogger(__name__)
+
+TIMEOUT = 15 * 60
+
+MYSQL_APP_NAME = "mysql-k8s"
+TEST_APP_NAME = "mysql-test-app"
+
+
+@pytest.mark.group(1)
+@pytest.mark.abort_on_fail
+async def test_deploy_latest(ops_test: OpsTest) -> None:
+    """Simple test to ensure that the mysql and application charms get deployed."""
+    await asyncio.gather(
+        ops_test.model.deploy(
+            MYSQL_APP_NAME,
+            application_name=MYSQL_APP_NAME,
+            num_units=3,
+            channel="8.0/stable",
+            constraints="mem=1G",
+            # config={"profile": "testing"}, # config not available in 8.0/stable@r75
+        ),
+        ops_test.model.deploy(
+            TEST_APP_NAME,
+            application_name=TEST_APP_NAME,
+            num_units=1,
+            channel="latest/edge",
+        ),
+    )
+    await relate_mysql_and_application(ops_test, MYSQL_APP_NAME, TEST_APP_NAME)
+    logger.info("Wait for applications to become active")
+    await ops_test.model.wait_for_idle(
+        apps=[MYSQL_APP_NAME, TEST_APP_NAME],
+        status="active",
+        timeout=TIMEOUT,
+    )
+    assert len(ops_test.model.applications[MYSQL_APP_NAME].units) == 3
+
+
+@pytest.mark.group(1)
+@pytest.mark.abort_on_fail
+async def test_upgrade_from_stable(ops_test: OpsTest):
+    """Test updating from stable channel."""
+    application = ops_test.model.applications[MYSQL_APP_NAME]
+    logger.info("Build charm locally")
+    charm = await ops_test.build_charm(".")
+
+    logger.info("Refresh the charm")
+    await application.refresh(path=charm)
+
+    logger.info("Wait for upgrade to start")
+    await ops_test.model.block_until(
+        lambda: "maintenance" in {unit.workload_status for unit in application.units},
+        timeout=TIMEOUT,
+    )
+
+    logger.info("Wait for upgrade to complete")
+    await ops_test.model.wait_for_idle(
+        apps=[MYSQL_APP_NAME], status="active", idle_period=30, timeout=TIMEOUT
+    )
+
+    logger.info("Ensure continuous_writes")
+    await ensure_all_units_continuous_writes_incrementing(ops_test)
